* Fixed: Possible runtime assert
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@9018 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/sdk/virtualbuildtargetsdlg.cpp
@@ -73,7 +73,7 @@ void VirtualBuildTargetsDlg::SetVirtualTarget(const wxString& targetName)
         return;
 
     wxArrayString checked;
-    for (int i = 0; i < (int)lstTargets->GetCount(); ++i)
+    for (unsigned int i = 0; i < lstTargets->GetCount(); ++i)
     {
         if (lstTargets->IsChecked(i))
             checked.Add(lstTargets->GetString(i));
